Pin : move some classes around and introduce PinCodeManager
