+++ features/pin/impl/build.gradle.kts
@@ -39,6 +39,7 @@ dependencies {
     implementation(projects.libraries.matrixui)
     implementation(projects.libraries.designsystem)
     implementation(projects.libraries.featureflag.api)
+    implementation(projects.libraries.cryptography.api)
 
     testImplementation(libs.test.junit)
     testImplementation(libs.coroutines.test)
